+++ test/somnus.spec.ts
@@ -1,4 +1,4 @@
-import { ISomnus, IRouteConfig } from "../src/somnus.d";
+import { ISomnus, IRouteConfig } from '../src/somnus.d';
 
 // this is actually declared in `setup.js` that runs before all tests
 declare const somnus: ISomnus;
@@ -15,9 +15,9 @@ describe('somnus', () => {
     somnus.stop(done);
   });
 
-  describe('without routeConfig', () => {
+  describe('without `routeConfig`', () => {
 
-    before((done) => {
+    beforeEach((done) => {
       somnus.start((addr): void => {
         currentPort = addr.port;
         done();
@@ -30,29 +30,67 @@ describe('somnus', () => {
 
   });
 
-  describe('with routeConfig', () => {
+  describe('with `routeConfig`', () => {
 
     const routeConfig: IRouteConfig = {
-      'get /hello': [
+      'get /echo': [
         (req, res, next) => next(),
-        (req, res, next) => res.send('world')
-      ]
+        (req, res, next) => res.send('echo echo')
+      ],
+      'post /echo': [
+        (req, res, next) => next(),
+        (req, res, next) => res.send({ foo: req.params.foo })
+      ],
+      'put /echo': (req, res) => res.send({ message: req.params })
     };
 
-    before((done) => {
+    beforeEach((done) => {
       somnus.start({ routeConfig }, (addr): void => {
         currentPort = addr.port;
         done();
       });
     });
 
+    afterEach(() => {
+      const { routes } = somnus.server.getDebugInfo();
+      for (const { name } of routes) {
+        somnus.server.rm(name);
+      }
+    });
+
     it('should respond as configured', (done) => {
-      fetch(`http://127.0.0.1:${currentPort}/hello`, {
+      fetch(`http://127.0.0.1:${currentPort}/echo`, {
         headers: { 'Accept': 'application/json' }
       })
         .then((val: Response) => val.json())
         .then((val: string) => {
-          assert(val === 'world');
+          assert(val === 'echo echo');
+          done();
+        });
+    });
+
+    it('should respond as configured', (done) => {
+      fetch(`http://127.0.0.1:${currentPort}/echo`, {
+        method: 'PUT',
+        headers: { 'Content-Type': 'application/json' },
+        body: JSON.stringify({ fooz: 'bars' })
+      })
+        .then((resp: Response) => resp.json())
+        .then((resp: { message: any }) => {
+          assert.strictEqual(resp.message.fooz, 'bars');
+          done();
+        });
+    });
+
+    it('should use the `bodyParser` plugin', (done) => {
+      fetch(`http://127.0.0.1:${currentPort}/echo`, {
+        method: 'POST',
+        headers: { 'Content-Type': 'application/json' },
+        body: JSON.stringify({ foo: 'bar' })
+      })
+        .then((resp: Response) => resp.json())
+        .then((resp: { foo: unknown }) => {
+          assert(resp.foo === 'bar');
           done();
         });
     });